Added user info card display on hover
This addresses the request to add a display for user info card on hover,
[issue #64](https://github.com/dubeaud/bugnet/issues/64).

I only added it to the Assigned User on the Issue List and the Creator
User for a comment. If it needs to be added elsewhere, feel free, but
this should be a step in the right direction.